backend/backend-service-for-prediction-clinker-free-lime/requirements.txt
google-cloud-aiplatform
Flask
Flask-Cors
gunicorn